feat(Navbar): add useGetUserProfile hook to fetch user profile data
The Navbar component now imports the useGetUserProfile hook from '@/hooks/useGetUserProfile' to fetch the user profile data. This hook will be used to display the user's avatar in the Navbar component.

---

feat(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics
feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port

In the server.ts file, the port variable has been changed from lowercase "port" to uppercase "PORT" to improve semantics. This change makes it clear that the variable represents the port number for the server.

Additionally, support for the process.env.PORT environment variable has been added. This allows the app to be run on a configurable port specified in the environment variables. This change provides flexibility in deploying the app to different environments.

---

feat(ProfilePage): add file upload functionality and update user profile

The ProfilePage component now includes functionality to upload a file (image) and update the user's profile. The component imports the useGetUserProfile hook from '@/hooks/useGetUserProfile' to fetch the user profile data.

The component includes a form with inputs for phone and document fields. When the form is submitted, the file upload functionality is triggered. The selected file is uploaded to the server using the /file-upload API endpoint. The uploaded file is associated with the user's profile and the profile data is updated using the /users/profile API endpoint.

The form submission triggers a POST request to the /file-upload and /users/profile API endpoints. The file upload request includes the file name and file type, and the profile update request includes the phone, document, and image (file key) fields.

After the profile is updated, the useGetUserProfile hook is used to refetch the user profile data and update the UI. The component also displays a success toast message when the profile is successfully updated.

---

fix(next.config.js): add new domain for image hosting

The next.config.js file has been updated to include a new domain for image hosting. The domain 'recogidas-y-despachos-2023-2-lw-media.s3.amazonaws.com' has been added to the 'domains' array in the images configuration. This allows images from this domain to be used in the application.

---

feat(package.json): add aws-sdk dependency

The package.json file has been updated to include the aws-sdk dependency